+++ ichnaea/tests/base.py
@@ -29,7 +29,6 @@ from ichnaea.async.config import (
 from ichnaea.cache import configure_redis
 from ichnaea.config import DummyConfig
 from ichnaea.constants import GEOIP_CITY_ACCURACY
-from ichnaea.country import country_max_radius
 from ichnaea.db import configure_db
 from ichnaea.geoip import configure_geoip
 from ichnaea.http import configure_http_session
@@ -40,6 +39,7 @@ from ichnaea.log import (
     DebugStatsClient,
 )
 from ichnaea.models import _Model, ApiKey
+from ichnaea.region import region_max_radius
 from ichnaea.webapp.config import main
 
 # make new unittest API's available under Python 2.6
@@ -106,7 +106,7 @@ GEOIP_DATA = {
         'ip': '67.43.156.1',
         'latitude': 27.5,
         'longitude': 90.5,
-        'accuracy': country_max_radius('BT'),
+        'accuracy': region_max_radius('BT'),
     },
 }